+++ application/src/main/java/run/halo/app/theme/finders/CommentPublicQueryService.java
@@ -1,0 +1,32 @@
+package run.halo.app.theme.finders;
+
+import java.util.Comparator;
+import org.springframework.lang.Nullable;
+import reactor.core.publisher.Mono;
+import run.halo.app.core.extension.content.Comment;
+import run.halo.app.core.extension.content.Reply;
+import run.halo.app.extension.ListResult;
+import run.halo.app.extension.Ref;
+import run.halo.app.theme.finders.vo.CommentVo;
+import run.halo.app.theme.finders.vo.ReplyVo;
+
+/**
+ * comment finder.
+ *
+ * @author LIlGG
+ */
+public interface CommentPublicQueryService {
+    Mono<CommentVo> getByName(String name);
+
+    Mono<ListResult<CommentVo>> list(Ref ref, @Nullable Integer page,
+        @Nullable Integer size);
+
+    Mono<ListResult<CommentVo>> list(Ref ref, @Nullable Integer page,
+        @Nullable Integer size, @Nullable Comparator<Comment> comparator);
+
+    Mono<ListResult<ReplyVo>> listReply(String commentName, @Nullable Integer page,
+        @Nullable Integer size);
+
+    Mono<ListResult<ReplyVo>> listReply(String commentName, @Nullable Integer page,
+        @Nullable Integer size, @Nullable Comparator<Reply> comparator);
+}
